Fix duration_min type to support decimal values
- Changed duration_min from INTEGER to NUMERIC(10, 1) in PostgreSQL
- Created migration 009 to update existing database
- This fixes "invalid input syntax for type integer: \"2.2\"" error
- Focus periods can now store fractional minute values (e.g., 2.2 minutes)

+++ Arzu_simulator_back/src/database/init-postgres.sql
@@ -72,7 +72,7 @@ CREATE TABLE IF NOT EXISTS focus_periods (
   start_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   end_time TIMESTAMP,
   duration INTEGER,
-  duration_min INTEGER,
+  duration_min NUMERIC(10, 1),
   is_interrupted BOOLEAN DEFAULT false,
   completed BOOLEAN DEFAULT false,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
